Random (router or modem) reset procedure added

+++ pingtest.sh
@@ -8,5 +8,13 @@ do
 	fi
 	tries=$((tries+1))
 done
+
 date >> /mnt/share/restarts.txt
-sleep 70 && touch /etc/banner && reboot
+sleep 1 && umount /mnt/share/
+
+r=$((0x$(cut -c1-1 /proc/sys/kernel/random/uuid) % 2))
+if [ $r -eq 1 ]; then
+ sleep 70 && touch /etc/banner && reboot
+else
+ sleep 5 && echo -e 'AT^RESET\r' > /dev/ttyUSB2
+fi
